reproduction/external/speculative_action_minimal/os-tuning/requirements.txt
google-genai
matplotlib
pytest
